+++ test/convert/explode/in/book.tf
@@ -1,0 +1,14 @@
+@node
+@valueType=str
+
+100	Genesis
+Exodus
+Leviticus
+Numeri
+Deuteronomium
+200	Josua
+Judices
+Samuel_I
+Samuel_II
+Reges_I
+Reges_II