+++ test/convert/explode/in/crossref.tf
@@ -1,0 +1,8 @@
+@edge
+@edgeValues
+@valueType=int
+
+1	5-10	10
+5-7,8	20
+10	12,14	75
+13,15	80