+++ test/convert/explode/in/mother.tf
@@ -1,0 +1,8 @@
+@edge
+@edgeValues
+@valueType=str
+
+1	5-10	apo
+5-7,8	epi
+10	12,14	dia
+13,15	pros