+++ test/convert/explode/in/otext.tf
@@ -1,0 +1,4 @@
+@config
+@fmt:text-orig-full={qere_utf8/g_word_utf8}{qere_trailer_utf8/trailer_utf8}
+@sectionFeatures=book,chapter,verse
+@sectionTypes=book,chapter,verse