+++ test/convert/explode/in/otype.tf
@@ -1,0 +1,7 @@
+@node
+@valueType=str
+
+1-3,5	word
+6-8	phrase
+9,10	clause
+11	sentence